requirements.txt
asgiref==3.5.2
astroid==2.12.13
colorama==0.4.6
dill==0.3.6
Django==3.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
platformdirs==2.5.4
PyJWT==2.6.0
pylint==2.15.6
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
wrapt==1.14.1
